refactor: Remove old synchronisation code
