--- coordinator/src/block_streams/mod.rs
@@ -1,5 +1,0 @@
-mod handler;
-mod synchronise;
-
-pub use handler::BlockStreamsHandler;
-pub use synchronise::synchronise_block_streams;
