--- coordinator/src/executors/mod.rs
@@ -1,5 +1,0 @@
-mod handler;
-mod synchronise;
-
-pub use handler::ExecutorsHandler;
-pub use synchronise::synchronise_executors;
