+++ coordinator/src/main.rs
@@ -5,15 +5,15 @@ use near_primitives::types::AccountId;
 use tokio::time::sleep;
 use tracing_subscriber::prelude::*;
 
-use crate::block_streams::BlockStreamsHandler;
-use crate::executors::ExecutorsHandler;
+use crate::block_streams_handler::BlockStreamsHandler;
+use crate::executors_handler::ExecutorsHandler;
 use crate::indexer_state::IndexerStateManager;
 use crate::redis::RedisClient;
 use crate::registry::Registry;
 use crate::synchroniser::Synchroniser;
 
-mod block_streams;
-mod executors;
+mod block_streams_handler;
+mod executors_handler;
 mod indexer_config;
 mod indexer_state;
 mod redis;
